remove assert version
